+++ src/main/scala/altitude/dao/mongo/StatDao.scala
@@ -6,7 +6,7 @@ import com.mongodb.casbah.Imports._
 
 class StatDao(val app: Altitude) extends BaseMongoDao("altitude_stats") with altitude.dao.StatDao {
 
-  def incrementStat(statName: String, count: Int = 1)(implicit txId: TransactionId): Unit = {
+  def incrementStat(statName: String, count: Long = 1)(implicit txId: TransactionId): Unit = {
     val query: DBObject =  MongoDBObject(
       C("Stat.DIMENSION") -> statName
     )
